backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.4.0
qdrant-client>=1.6.0
openai>=1.3.0
sqlalchemy>=2.0.23
python-multipart>=0.0.6
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
httpx>=0.25.0
